Implement AbstractUpdater.
