+++ src/Prometheus/Storage/RedisTxn/Updater/SummaryUpdater.php
@@ -4,10 +4,8 @@ namespace Prometheus\Storage\RedisTxn\Updater;
 
 use Prometheus\Storage\RedisTxn\Metric\MetadataBuilder;
 use Prometheus\Storage\RedisTxn\RedisScript\RedisScript;
-use Prometheus\Storage\RedisTxn\RedisScript\RedisScriptHelper;
-use Redis;
 
-class SummaryUpdater implements UpdaterInterface
+class SummaryUpdater extends AbstractUpdater
 {
     /**
      * @var string
@@ -33,39 +31,6 @@ redis.call('zadd', summaryKey, currentTime, summaryValue)
 redis.call('hset', metaHashKey, summaryKey, summaryMetadata)
 redis.call('hset', metaHashKey, ttlFieldName, ttlFieldValue)
 LUA;
-
-    /**
-     * @var RedisScriptHelper
-     */
-    private $helper;
-
-    /**
-     * @var Redis
-     */
-    private $redis;
-
-    /**
-     * @param Redis $redis
-     */
-    public function __construct(Redis $redis)
-    {
-        $this->helper = new RedisScriptHelper();
-        $this->redis = $redis;
-    }
-
-    public function getHelper(): RedisScriptHelper
-    {
-        $this->helper = $this->helper ?? new RedisScriptHelper();
-        return $this->helper;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getRedis(): Redis
-    {
-        return $this->redis;
-    }
 
     /**
      * @inheritDoc
@@ -111,13 +76,5 @@ LUA;
             ->withArgs($scriptArgs)
             ->withNumKeys(3)
             ->build();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function update(array $data)
-    {
-        return $this->getRedisScript($data)->eval($this->redis);
     }
 }
